Moved the factory to the correct directory
