+++ src/Factories/OperationFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App;
+namespace App\Factories;
 
 use App\Enums\OperatorEnum;
 use App\Operations\Division;
